fix(history): Handle intermediate expiry states
The current model is to have 48 hours of gateway validity with an
additional 24 hours for minting. The current implementation will
show an expiry error in the 24 hours minting grace period, breaking
the flow.

This fixes that by only showing the expiry warning when there are
no detected deposits for the given gateway.

It still has an unfortunate state where the "semi-expired" transactions
show as pending, but we will leave this for fixing after release